moved styles to styles.js
