--- src/app/domains/Condition/components/ConditionForm/ConditionTemplates/index.js
@@ -1,1 +1,0 @@
-// export { ChoiceTemplate, PlaneTextStringTemplate }
