+++ src/app/domains/Condition/components/ConditionTemplates/FileUploadTemplate/index.js
