+++ src/app/domains/Condition/components/ConditionTemplates/OpinionScaleTemplate/index.js
