+++ src/app/domains/Condition/components/ConditionTemplates/PictureChoiceTemplate/index.js
