+++ src/app/domains/Condition/components/ConditionTemplates/PlainLongTextStringTemplate/PlainLongTextStringTemplate.style.js
@@ -1,0 +1,4 @@
+export const styles = {
+  bgc: { backgroundColor: '#d6e1f2' },
+  fontColor: { color: '#1d6fdc' }
+}
