+++ src/app/domains/Condition/components/ConditionTemplates/PlainLongTextStringTemplate/index.js
