+++ src/app/domains/Condition/components/ConditionTemplates/PlainShortTextStringTemplate/index.js
