+++ src/app/domains/Condition/components/ConditionTemplates/PlainTextDateTemplate/index.js
