+++ src/app/domains/Condition/components/ConditionTemplates/RatingTemplate/RatingTemplate.js
@@ -4,12 +4,11 @@ import Text from 'antd/lib/typography/Text'
 import React from 'react'
 import { QuestionSelect } from 'domains/Question/components'
 import PropTypes from 'prop-types'
-import PlaneLongTextStringTemplate from 'domains/Condition/components/ConditionForm/ConditionTemplates/PlainLongTextStringTemplate'
+import { styles } from './RatingTemplate.styles'
 
-const { Option, OptGroup } = Select
-
-function PictureChoiceTemplate(props) {
+function RatingTemplate(props) {
   const { answers, questionList, addRedirectQuestion } = props
+
   let startLetter = 65
 
   return (
@@ -20,11 +19,12 @@ function PictureChoiceTemplate(props) {
             <Box
               display="flex"
               alignItems="center"
+              key={index}
               p={2}
               mr={2}
               border="1px solid #bbbbbb"
               borderRadius="4px">
-              <Button type="outline" style={{ marginRight: '10px' }}>
+              <Button type="outline" style={styles.buttonM}>
                 <Text strong>{String.fromCharCode(startLetter++)}</Text>
               </Button>
               {item.name}
@@ -43,9 +43,9 @@ function PictureChoiceTemplate(props) {
     </>
   )
 }
-PictureChoiceTemplate.propTypes = {
+RatingTemplate.propTypes = {
   answers: PropTypes.array,
   questionList: PropTypes.array,
   addRedirectQuestion: PropTypes.func
 }
-export default PictureChoiceTemplate
+export default RatingTemplate
