+++ src/app/domains/Condition/components/ConditionTemplates/RatingTemplate/index.js
