+++ src/app/domains/Condition/components/ConditionTemplates/YesNoChoiceTemplate/YesNoChoiceTemplate.styles.js
@@ -1,0 +1,3 @@
+export const styles = {
+  buttonM: { marginRight: '10px' }
+}
