+++ src/app/domains/Condition/components/ConditionTemplates/YesNoChoiceTemplate/index.js
