+++ src/app/domains/Condition/components/ConditionTemplates/index.js
@@ -1,0 +1,8 @@
+export { default as FileUploadTemplate } from './FileUploadTemplate'
+export { default as OpinionScaleTemplate } from './OpinionScaleTemplate'
+export { default as PictureChoiceTemplate } from './PictureChoiceTemplate'
+export { default as PlaneLongTextStringTemplate } from './PlainLongTextStringTemplate'
+export { default as PlaneShortTextStringTemplate } from './PlainShortTextStringTemplate'
+export { default as PlaneTextDateTemplate } from './PlainTextDateTemplate'
+export { default as RatingTemplate } from './RatingTemplate'
+export { default as YesNoChoiceTemplate } from './YesNoChoiceTemplate'
